# src/app/api/openapi.yaml
openapi: 3.0.3
info:
  title: Dupip Morpheus API
  version: 1.0.0
  description: >
    HTTP API for the Dupip application. Authentication uses Clerk session
    cookies; most routes require an authenticated user. Errors use the shape
    `{ "error": "message" }`.
servers:
  - url: https://www.dupip.com
    description: Production
  - url: https://beta.dupip.com
    description: Nightly
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /api/v1/jobs/{jobId}:
    get:
      tags: [jobs]
      summary: Get a job
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Job }
    put:
      tags: [jobs]
      summary: Update a job (status transitions, reviews, notes, justification, evidence documentIds, location)
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Updated }
    delete:
      tags: [jobs]
      summary: Cancel a job
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Cancelled }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __session